refactor(net): keep guard profile builder internal
